Fix bank switching problem for tile : when a tile gets a bank inferior to its current one => display wasn't updated. Because tile are displayed with ascendant order for bank.

+++ trunk/zildo/src/zildo/monde/map/Area.java
@@ -222,7 +222,7 @@ public class Area implements EasySerializable {
 		}
 		Tile back = temp.getBackTile();
 		back.index = quoi & 255;
-		back.bank = quoi >> 8;
+		back.bank = (byte) (quoi >> 8);
 		this.set_mapcase(x, y + 4, temp);
 
 		changes.add(new Point(x, y + 4));
